change protorun to generate
